requirements.txt
clint==0.3.1
coverage==3.6
docopt==0.6.1
flake8==2.0
watchdog==0.6.0
PyYAML==3.10
